Implemented pagination, sorting and dynamic filtering using JPA Specifications

+++ src/main/java/com/shahid/shopsphere/service/imp1/CartServiceImpl.java
@@ -1,24 +1,24 @@
 package com.shahid.shopsphere.service.imp1;
 
 import java.math.BigDecimal;
-import java.util.List;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.shahid.shopsphere.dto.cart.AddToCartRequest;
-import com.shahid.shopsphere.dto.cart.CartItemResponse;
 import com.shahid.shopsphere.dto.cart.CartResponse;
 import com.shahid.shopsphere.dto.cart.UpdateCartItemRequest;
 import com.shahid.shopsphere.entity.Cart;
 import com.shahid.shopsphere.entity.CartItem;
 import com.shahid.shopsphere.entity.Product;
 import com.shahid.shopsphere.entity.User;
+import com.shahid.shopsphere.exception.ForbiddenException;
 import com.shahid.shopsphere.exception.ResourceNotFoundException;
+import com.shahid.shopsphere.mapper.CartMapper;
 import com.shahid.shopsphere.repository.CartItemsRepository;
 import com.shahid.shopsphere.repository.CartRepository;
 import com.shahid.shopsphere.repository.ProductRepository;
-import com.shahid.shopsphere.repository.UserRepository;
+import com.shahid.shopsphere.service.AuthenticationFacade;
 import com.shahid.shopsphere.service.CartService;
 
 import lombok.RequiredArgsConstructor;
@@ -31,28 +31,21 @@ public class CartServiceImpl implements CartService {
     private final CartRepository cartRepository;
     private final CartItemsRepository cartItemsRepository;
     private final ProductRepository productRepository;
-    private final UserRepository userRepository;
-     
-     private CartResponse mapToCartResponse(Cart cart) {
+    private final AuthenticationFacade authenticationFacade;
+    private final CartMapper cartMapper;
+      /**
+     * Calculates total amount of the cart.
+     */
+    private BigDecimal calculateCartTotal(Long cartId) {
 
-        List<CartItemResponse> items = cart.getCartItems()
+        return cartItemsRepository.findByCartId(cartId)
                 .stream()
-                .map(item -> new CartItemResponse(
-                        item.getProduct().getId(),
-                        item.getProduct().getName(),
-                        item.getUnitPrice(),
-                        item.getQuantity(),
+                .map(item ->
                         item.getUnitPrice()
-                                .multiply(BigDecimal.valueOf(item.getQuantity()))
-                ))
-                .toList();
-
-        return new CartResponse(
-                cart.getId(),
-                items,
-                cart.getTotalAmount()
-        );
+                                .multiply(BigDecimal.valueOf(item.getQuantity())))
+                .reduce(BigDecimal.ZERO, BigDecimal::add);
     }
+
     @Override
     public CartResponse addToCart(AddToCartRequest request) {
 
@@ -62,10 +55,7 @@ public class CartServiceImpl implements CartService {
                         new ResourceNotFoundException("Product not found"));
 
         // 2. Get User (Temporary)
-        User user = userRepository.findById(1L)
-                .orElseThrow(() ->
-                        new ResourceNotFoundException("User not found"));
-
+        User user =authenticationFacade.getCurrentUser();
         // 3. Find or Create Cart
         Cart cart = cartRepository.findByUserId(user.getId())
                 .orElseGet(() -> {
@@ -100,19 +90,15 @@ public class CartServiceImpl implements CartService {
         cartItemsRepository.save(cartItem);
 
         // 6. Calculate Total Amount
-        BigDecimal total = cartItemsRepository.findByCartId(cart.getId())
-                .stream()
-                .map(item ->
-                        item.getUnitPrice()
-                                .multiply(BigDecimal.valueOf(item.getQuantity())))
-                .reduce(BigDecimal.ZERO, BigDecimal::add);
-
+        BigDecimal total = calculateCartTotal(cart.getId());
         // 7. Update Cart Total
         cart.setTotalAmount(total);
         cartRepository.save(cart);
-
+        //reload
+          cart = cartRepository.findById(cart.getId())
+        .orElseThrow(() -> new ResourceNotFoundException("Cart not found"));
         // 8. Return Response
-        return mapToCartResponse(cart);
+        return cartMapper.toCartResponse(cart);
     }
 
    
@@ -120,18 +106,23 @@ public class CartServiceImpl implements CartService {
 
     public CartResponse getCart(){
         //fetch user
-        User user = userRepository.findById(1L).orElseThrow(() -> new ResourceNotFoundException("user Not Found."));
+        User user = authenticationFacade.getCurrentUser();
 
         //fetch cart of user
         Cart cart = cartRepository.findByUserId(user.getId()).orElseThrow(()-> new ResourceNotFoundException("cart not Found."));
 
-        return mapToCartResponse(cart);
+        return cartMapper.toCartResponse(cart);
     }
 
     //update cart item
     public CartResponse updateCartItem(Long Itemid,UpdateCartItemRequest request)
     {          //find that item id incart
              CartItem cartItem = cartItemsRepository.findById(Itemid).orElseThrow(() -> new ResourceNotFoundException("Cart Item Not Found."));
+             User currentUser = authenticationFacade.getCurrentUser();
+
+if (!cartItem.getCart().getUser().getId().equals(currentUser.getId())) {
+    throw new ForbiddenException("You cannot modify another user's cart.");
+}
              //update quantity
              cartItem.setQuantity(request.getQuantity());
              //save 
@@ -139,15 +130,11 @@ public class CartServiceImpl implements CartService {
              //getCart
              Cart cart = cartItem.getCart();
              //toatl amount
-             BigDecimal total = cartItemsRepository.findByCartId(cart.getId())
-                                                    .stream()
-                                                     .map(item -> item.getUnitPrice().multiply(BigDecimal.valueOf(item.getQuantity())))
-                                                     .reduce(BigDecimal.ZERO,BigDecimal::add);
-                                                     
+             BigDecimal total = calculateCartTotal(cart.getId());
           //save total amount
           cart.setTotalAmount(total);
           cartRepository.save(cart);
-          return mapToCartResponse(cart);
+          return cartMapper.toCartResponse(cart);
 
     }
 
@@ -155,30 +142,31 @@ public class CartServiceImpl implements CartService {
 public CartResponse removeCartItem(Long itemId) 
 {         //find cartitem
         CartItem cartItem = cartItemsRepository.findById(itemId).orElseThrow(() -> new ResourceNotFoundException("Cart Not Found."));
-           //remove
+          User currentUser = authenticationFacade.getCurrentUser();
+
+if (!cartItem.getCart().getUser().getId().equals(currentUser.getId())) {
+    throw new ForbiddenException("You cannot modify another user's cart.");
+}  
+        //remove
            cartItemsRepository.delete(cartItem);
 
            //get cart
            Cart cart =cartItem.getCart();
            //update total amount
 
-           BigDecimal total = cartItemsRepository.findByCartId(cart.getId())
-                                                  .stream()
-                                                 .map(item -> item.getUnitPrice().multiply(BigDecimal.valueOf(item.getQuantity())))
-                                                .reduce(BigDecimal.ZERO,BigDecimal::add);
+           BigDecimal total = calculateCartTotal(cart.getId());
 
            cart.setTotalAmount(total);
            cartRepository.save(cart);
-           return mapToCartResponse(cart);
+           return cartMapper.toCartResponse(cart);
 
 }
 public void clearCart(){
 
-        User user = userRepository.findById(1L).orElseThrow(() -> new ResourceNotFoundException("User Not Found."));
+        User user = authenticationFacade.getCurrentUser();
         Cart cart = cartRepository.findByUserId(user.getId()).orElseThrow(() -> new ResourceNotFoundException("Cart Not Found."));
-
         cartItemsRepository.deleteAll(cart.getCartItems());
-
+        cart.getCartItems().clear();
         cart.setTotalAmount(BigDecimal.ZERO);
         cartRepository.save(cart);
